refactore: 重构trap模块
